requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.7
Django==2.0
django-appconf==1.0.4
django-cryptography==1.0
django-fernet-fields==0.6
django-widget-tweaks==1.4.1
djangorestframework==3.7.3
gunicorn==20.1.0
idna==2.6
pycparser==2.20
pytz==2017.3
requests==2.18.4
sqlparse==0.4.1
urllib3==1.22
waitress==2.0.0
whitenoise==5.2.0
